@@ -1,12 +1,12 @@
 name: connectycube_flutter_call_kit
 description: A Flutter plugin for displaying call screen when app in background or terminated.
-version: 0.0.1-dev.1
+version: 0.1.0-dev.1
 homepage: https://github.com/ConnectyCube/connectycube-flutter-call-kit
 issue_tracker: https://github.com/ConnectyCube/connectycube-flutter-call-kit/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: '>=1.20.0'
 
 dependencies:
   flutter:
