@@ -1,7 +1,7 @@
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io';
 
-import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 typedef Future<dynamic> CallEventHandler(
@@ -10,6 +10,7 @@ typedef Future<dynamic> CallEventHandler(
   int callerId,
   String callerName,
   Set<int> opponentsIds,
+  Map<String, String>? userInfo,
 );
 
 class ConnectycubeFlutterCallKit {
@@ -17,14 +18,14 @@ class ConnectycubeFlutterCallKit {
       const MethodChannel('connectycube_flutter_call_kit');
 
   static ConnectycubeFlutterCallKit get instance => _getInstance();
-  static ConnectycubeFlutterCallKit _instance;
+  static ConnectycubeFlutterCallKit? _instance;
   static String TAG = "ConnectycubeFlutterCallKit";
 
   static ConnectycubeFlutterCallKit _getInstance() {
     if (_instance == null) {
       _instance = ConnectycubeFlutterCallKit._internal();
     }
-    return _instance;
+    return _instance!;
   }
 
   factory ConnectycubeFlutterCallKit() => _getInstance();
@@ -37,15 +38,24 @@ class ConnectycubeFlutterCallKit {
     int callerId,
     String callerName,
     Set<int> opponentsIds,
-  ) onCallAcceptedWhenTerminated;
+    Map<String, String>? userInfo,
+  )? onCallRejectedWhenTerminated;
 
-  static CallEventHandler _onCallAccepted;
-  static CallEventHandler _onCallRejected;
+  static Function(
+    String sessionId,
+    int callType,
+    int callerId,
+    String callerName,
+    Set<int> opponentsIds,
+    Map<String, String>? userInfo,
+  )? onCallAcceptedWhenTerminated;
 
-  /// Sets up [MessageHandler] for incoming messages.
+  static CallEventHandler? _onCallAccepted;
+  static CallEventHandler? _onCallRejected;
+
   void init({
-    CallEventHandler onCallAccepted,
-    CallEventHandler onCallRejected,
+    CallEventHandler? onCallAccepted,
+    CallEventHandler? onCallRejected,
   }) {
     _onCallAccepted = onCallAccepted;
     _onCallRejected = onCallRejected;
@@ -57,11 +67,12 @@ class ConnectycubeFlutterCallKit {
   }
 
   static Future<void> showCallNotification({
-    @required String sessionId,
-    @required int callType,
-    @required int callerId,
-    @required String callerName,
-    @required Set<int> opponentsIds,
+    required String? sessionId,
+    required int? callType,
+    required int? callerId,
+    required String? callerName,
+    required Set<int>? opponentsIds,
+    Map<String, String>? userInfo,
   }) async {
     if (!Platform.isAndroid) return;
 
@@ -70,12 +81,13 @@ class ConnectycubeFlutterCallKit {
       'call_type': callType,
       'caller_id': callerId,
       'caller_name': callerName,
-      'call_opponents': opponentsIds.join(','),
+      'call_opponents': opponentsIds!.join(','),
+      'user_info': userInfo != null ? jsonEncode(userInfo) : jsonEncode(Map()),
     });
   }
 
   static Future<void> reportCallAccepted({
-    @required String sessionId,
+    required String? sessionId,
   }) async {
     if (!Platform.isAndroid) return;
 
@@ -85,7 +97,7 @@ class ConnectycubeFlutterCallKit {
   }
 
   static Future<void> reportCallEnded({
-    @required String sessionId,
+    required String? sessionId,
   }) async {
     if (!Platform.isAndroid) return;
 
@@ -95,18 +107,20 @@ class ConnectycubeFlutterCallKit {
   }
 
   static Future<String> getCallState({
-    @required String sessionId,
+    required String? sessionId,
   }) async {
     if (!Platform.isAndroid) return Future.value(CallState.UNKNOWN);
 
     return _channel.invokeMethod("getCallState", {
       'session_id': sessionId,
+    }).then((state) {
+      return state.toString();
     });
   }
 
   static Future<void> setCallState({
-    @required String sessionId,
-    @required String callState,
+    required String? sessionId,
+    required String? callState,
   }) async {
     return _channel.invokeMethod("setCallState", {
       'session_id': sessionId,
@@ -114,8 +128,39 @@ class ConnectycubeFlutterCallKit {
     });
   }
 
+  static Future<Map<String, dynamic>?> getCallData({
+    required String? sessionId,
+  }) async {
+    if (!Platform.isAndroid) return Future.value();
+
+    return _channel.invokeMethod("getCallData", {
+      'session_id': sessionId,
+    }).then((data) {
+      if (data == null) {
+        return Future.value(null);
+      }
+      return Future.value(Map<String, dynamic>.from(data));
+    });
+  }
+
+  static Future<void> clearCallData({
+    required String? sessionId,
+  }) async {
+    if (!Platform.isAndroid) return Future.value();
+
+    return _channel.invokeMethod("clearCallData", {
+      'session_id': sessionId,
+    });
+  }
+
+  static Future<String?> getLastCallId() async {
+    if (!Platform.isAndroid) return Future.value();
+
+    return _channel.invokeMethod("getLastCallId");
+  }
+
   static Future<void> setOnLockScreenVisibility({
-    @required bool isVisible,
+    required bool? isVisible,
   }) async {
     if (!Platform.isAndroid) return;
 
@@ -124,23 +169,17 @@ class ConnectycubeFlutterCallKit {
     });
   }
 
-  static Future<void> _handleMethod(MethodCall call) {
+  static Future<dynamic> _handleMethod(MethodCall call) {
     final Map map = call.arguments.cast<String, dynamic>();
     switch (call.method) {
       case "onCallAccepted":
-        return _onCallAccepted(
-          map["session_id"],
-          map["call_type"],
-          map["caller_id"],
-          map["caller_name"],
-          (map["call_opponents"] as String)
-              .split(',')
-              .map((stringUserId) => int.parse(stringUserId))
-              .toSet(),
-        );
-      case "onCallRejected":
+        var userInfo = map['user_info'];
+        var userInfoParsed;
+        if (userInfo != null) {
+          userInfoParsed = Map<String, String>.from(jsonDecode(userInfo));
+        }
         if (onCallAcceptedWhenTerminated != null) {
-          onCallAcceptedWhenTerminated.call(
+          onCallAcceptedWhenTerminated!.call(
             map["session_id"],
             map["call_type"],
             map["caller_id"],
@@ -149,21 +188,66 @@ class ConnectycubeFlutterCallKit {
                 .split(',')
                 .map((stringUserId) => int.parse(stringUserId))
                 .toSet(),
+            userInfoParsed,
           );
         }
-        return _onCallRejected(
-          map["session_id"],
-          map["call_type"],
-          map["caller_id"],
-          map["caller_name"],
-          (map["call_opponents"] as String)
-              .split(',')
-              .map((stringUserId) => int.parse(stringUserId))
-              .toSet(),
-        );
+
+        if (_onCallAccepted != null) {
+          return _onCallAccepted!(
+            map["session_id"],
+            map["call_type"],
+            map["caller_id"],
+            map["caller_name"],
+            (map["call_opponents"] as String)
+                .split(',')
+                .map((stringUserId) => int.parse(stringUserId))
+                .toSet(),
+            userInfoParsed,
+          );
+        }
+
+        break;
+      case "onCallRejected":
+        var userInfo = map['user_info'];
+        var userInfoParsed;
+        if (userInfo != null) {
+          userInfoParsed = Map<String, String>.from(jsonDecode(userInfo));
+        }
+
+        if (onCallRejectedWhenTerminated != null) {
+          onCallRejectedWhenTerminated!.call(
+            map["session_id"],
+            map["call_type"],
+            map["caller_id"],
+            map["caller_name"],
+            (map["call_opponents"] as String)
+                .split(',')
+                .map((stringUserId) => int.parse(stringUserId))
+                .toSet(),
+            userInfoParsed,
+          );
+        }
+
+        if (_onCallRejected != null) {
+          return _onCallRejected!(
+            map["session_id"],
+            map["call_type"],
+            map["caller_id"],
+            map["caller_name"],
+            (map["call_opponents"] as String)
+                .split(',')
+                .map((stringUserId) => int.parse(stringUserId))
+                .toSet(),
+            userInfoParsed,
+          );
+        }
+
+        break;
       default:
         throw UnsupportedError("Unrecognized JSON message");
     }
+
+    return Future.value();
   }
 }
 
