@@ -16,8 +16,10 @@ import androidx.annotation.Nullable
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
 
 
-fun createStartIncomingScreenIntent(context: Context, callId: String, callType: Int, callInitiatorId: Int,
-                                    callInitiatorName: String, opponents: ArrayList<Int>): Intent {
+fun createStartIncomingScreenIntent(
+    context: Context, callId: String, callType: Int, callInitiatorId: Int,
+    callInitiatorName: String, opponents: ArrayList<Int>, userInfo: String
+): Intent {
     val intent = Intent(context, IncomingCallActivity::class.java)
     intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
     intent.putExtra(EXTRA_CALL_ID, callId)
@@ -25,6 +27,7 @@ fun createStartIncomingScreenIntent(context: Context, callId: String, callType: 
     intent.putExtra(EXTRA_CALL_INITIATOR_ID, callInitiatorId)
     intent.putExtra(EXTRA_CALL_INITIATOR_NAME, callInitiatorName)
     intent.putIntegerArrayListExtra(EXTRA_CALL_OPPONENTS, opponents)
+    intent.putExtra(EXTRA_CALL_USER_INFO, userInfo)
     return intent
 }
 
@@ -37,6 +40,7 @@ class IncomingCallActivity : Activity() {
     private var callInitiatorId = -1
     private var callInitiatorName: String? = null
     private var callOpponents: ArrayList<Int>? = ArrayList()
+    private var callUserInfo: String? = null
 
 
     override fun onCreate(@Nullable savedInstanceState: Bundle?) {
@@ -48,8 +52,10 @@ class IncomingCallActivity : Activity() {
             setShowWhenLocked(true)
             setTurnScreenOn(true)
         } else {
-            window.addFlags(WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED or
-                    WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON)
+            window.addFlags(
+                WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED or
+                        WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON
+            )
         }
 
         processIncomingData(intent)
@@ -109,13 +115,17 @@ class IncomingCallActivity : Activity() {
         callInitiatorId = intent.getIntExtra(EXTRA_CALL_INITIATOR_ID, -1)
         callInitiatorName = intent.getStringExtra(EXTRA_CALL_INITIATOR_NAME)
         callOpponents = intent.getIntegerArrayListExtra(EXTRA_CALL_OPPONENTS)
+        callUserInfo = intent.getStringExtra(EXTRA_CALL_USER_INFO)
     }
 
     private fun initUi() {
-        val callTitleTxt: TextView = findViewById(resources.getIdentifier("user_name_txt", "id", packageName))
+        val callTitleTxt: TextView =
+            findViewById(resources.getIdentifier("user_name_txt", "id", packageName))
         callTitleTxt.text = callInitiatorName
-        val callSubTitleTxt: TextView = findViewById(resources.getIdentifier("call_type_txt", "id", packageName))
-        callSubTitleTxt.text = String.format(CALL_TYPE_PLACEHOLDER, if (callType == 1) "Video" else "Audio")
+        val callSubTitleTxt: TextView =
+            findViewById(resources.getIdentifier("call_type_txt", "id", packageName))
+        callSubTitleTxt.text =
+            String.format(CALL_TYPE_PLACEHOLDER, if (callType == 1) "Video" else "Audio")
     }
 
     // calls from layout file
@@ -126,6 +136,7 @@ class IncomingCallActivity : Activity() {
         bundle.putInt(EXTRA_CALL_INITIATOR_ID, callInitiatorId)
         bundle.putString(EXTRA_CALL_INITIATOR_NAME, callInitiatorName)
         bundle.putIntegerArrayList(EXTRA_CALL_OPPONENTS, callOpponents)
+        bundle.putString(EXTRA_CALL_USER_INFO, callUserInfo)
 
         val endCallIntent = Intent(this, EventReceiver::class.java)
         endCallIntent.action = ACTION_CALL_REJECT
@@ -141,6 +152,7 @@ class IncomingCallActivity : Activity() {
         bundle.putInt(EXTRA_CALL_INITIATOR_ID, callInitiatorId)
         bundle.putString(EXTRA_CALL_INITIATOR_NAME, callInitiatorName)
         bundle.putIntegerArrayList(EXTRA_CALL_OPPONENTS, callOpponents)
+        bundle.putString(EXTRA_CALL_USER_INFO, callUserInfo)
 
         val startCallIntent = Intent(this, EventReceiver::class.java)
         startCallIntent.action = ACTION_CALL_ACCEPT
