@@ -1,6 +1,5 @@
 package com.connectycube.flutter.connectycube_flutter_call_kit
 
-import android.R
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.app.PendingIntent
@@ -26,29 +25,76 @@ fun cancelCallNotification(context: Context, callId: String) {
     notificationManager.cancel(callId.hashCode())
 }
 
-fun showCallNotification(context: Context, callId: String, callType: Int, callInitiatorId: Int,
-                         callInitiatorName: String, callOpponents: ArrayList<Int>) {
+fun showCallNotification(
+    context: Context, callId: String, callType: Int, callInitiatorId: Int,
+    callInitiatorName: String, callOpponents: ArrayList<Int>, userInfo: String
+) {
     val notificationManager = NotificationManagerCompat.from(context)
 
     val intent = getLaunchIntent(context)
 
-    val pendingIntent = PendingIntent.getActivity(context, callId.hashCode(), intent, PendingIntent.FLAG_UPDATE_CURRENT)
+    val pendingIntent = PendingIntent.getActivity(
+        context,
+        callId.hashCode(),
+        intent,
+        PendingIntent.FLAG_UPDATE_CURRENT
+    )
 
-    val ringtone: Uri = RingtoneManager.getActualDefaultRingtoneUri(context.applicationContext, RingtoneManager.TYPE_RINGTONE)
+    val ringtone: Uri = RingtoneManager.getActualDefaultRingtoneUri(
+        context.applicationContext,
+        RingtoneManager.TYPE_RINGTONE
+    )
 
-    val callTypeTitle = String.format(CALL_TYPE_PLACEHOLDER, if (callType == 1) "Video" else "Audio")
+    val callTypeTitle =
+        String.format(CALL_TYPE_PLACEHOLDER, if (callType == 1) "Video" else "Audio")
 
-    val builder: NotificationCompat.Builder = createCallNotification(context, callInitiatorName, callTypeTitle, pendingIntent, ringtone)
+    val builder: NotificationCompat.Builder =
+        createCallNotification(context, callInitiatorName, callTypeTitle, pendingIntent, ringtone)
 
     // Add actions
-    addCallRejectAction(context, builder, callId, callType, callInitiatorId, callInitiatorName, callOpponents)
-    addCallAcceptAction(context, builder, callId, callType, callInitiatorId, callInitiatorName, callOpponents)
+    addCallRejectAction(
+        context,
+        builder,
+        callId,
+        callType,
+        callInitiatorId,
+        callInitiatorName,
+        callOpponents,
+        userInfo
+    )
+    addCallAcceptAction(
+        context,
+        builder,
+        callId,
+        callType,
+        callInitiatorId,
+        callInitiatorName,
+        callOpponents,
+        userInfo
+    )
 
     // Add full screen intent (to show on lock screen)
-    addCallFullScreenIntent(context, builder, callId, callType, callInitiatorId, callInitiatorName, callOpponents)
+    addCallFullScreenIntent(
+        context,
+        builder,
+        callId,
+        callType,
+        callInitiatorId,
+        callInitiatorName,
+        callOpponents,
+        userInfo
+    )
 
     // Add action when delete call notification
-    addCancelCallNotificationIntent(context, builder, callId, callType, callInitiatorId, callInitiatorName)
+    addCancelCallNotificationIntent(
+        context,
+        builder,
+        callId,
+        callType,
+        callInitiatorId,
+        callInitiatorName,
+        userInfo
+    )
 
     // Set small icon for notification
     setNotificationSmallIcon(context, builder)
@@ -67,114 +113,179 @@ fun getLaunchIntent(context: Context): Intent? {
     return packageManager.getLaunchIntentForPackage(packageName)
 }
 
-fun createCallNotification(context: Context, title: String, text: String?, pendingIntent: PendingIntent, ringtone: Uri): NotificationCompat.Builder {
+fun createCallNotification(
+    context: Context,
+    title: String,
+    text: String?,
+    pendingIntent: PendingIntent,
+    ringtone: Uri
+): NotificationCompat.Builder {
     val notificationBuilder = NotificationCompat.Builder(context, CALL_CHANNEL_ID)
     notificationBuilder
-            .setDefaults(NotificationCompat.DEFAULT_VIBRATE)
-            .setContentTitle(title)
-            .setContentText(text)
-            .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
-            .setAutoCancel(true)
-            .setOngoing(true)
-            .setCategory(NotificationCompat.CATEGORY_CALL)
-            .setContentIntent(pendingIntent)
-            .setSound(ringtone)
-            .setPriority(NotificationCompat.PRIORITY_MAX)
-            .setTimeoutAfter(60000)
+        .setDefaults(NotificationCompat.DEFAULT_VIBRATE)
+        .setContentTitle(title)
+        .setContentText(text)
+        .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
+        .setAutoCancel(true)
+        .setOngoing(true)
+        .setCategory(NotificationCompat.CATEGORY_CALL)
+        .setContentIntent(pendingIntent)
+        .setSound(ringtone)
+        .setPriority(NotificationCompat.PRIORITY_MAX)
+        .setTimeoutAfter(60000)
     return notificationBuilder
 }
 
-fun addCallRejectAction(context: Context, notificationBuilder: NotificationCompat.Builder,
-                        callId: String, callType: Int, callInitiatorId: Int, callInitiatorName: String, opponents: ArrayList<Int>) {
+fun addCallRejectAction(
+    context: Context,
+    notificationBuilder: NotificationCompat.Builder,
+    callId: String,
+    callType: Int,
+    callInitiatorId: Int,
+    callInitiatorName: String,
+    opponents: ArrayList<Int>,
+    userInfo: String
+) {
     val bundle = Bundle()
     bundle.putString(EXTRA_CALL_ID, callId)
     bundle.putInt(EXTRA_CALL_TYPE, callType)
     bundle.putInt(EXTRA_CALL_INITIATOR_ID, callInitiatorId)
     bundle.putString(EXTRA_CALL_INITIATOR_NAME, callInitiatorName)
     bundle.putIntegerArrayList(EXTRA_CALL_OPPONENTS, opponents)
+    bundle.putString(EXTRA_CALL_USER_INFO, userInfo)
 
     val declinePendingIntent: PendingIntent = PendingIntent.getBroadcast(
-            context,
-            callId.hashCode(),
-            Intent(context, EventReceiver::class.java)
-                    .setAction(ACTION_CALL_REJECT)
-                    .putExtras(bundle),
-            PendingIntent.FLAG_UPDATE_CURRENT)
+        context,
+        callId.hashCode(),
+        Intent(context, EventReceiver::class.java)
+            .setAction(ACTION_CALL_REJECT)
+            .putExtras(bundle),
+        PendingIntent.FLAG_UPDATE_CURRENT
+    )
     val declineAction: NotificationCompat.Action = NotificationCompat.Action.Builder(
-            R.drawable.ic_menu_close_clear_cancel,
-            getColorizedText("Reject", "#E02B00"),
-            declinePendingIntent)
-            .build()
+        context.resources.getIdentifier(
+            "ic_menu_close_clear_cancel",
+            "drawable",
+            context.packageName
+        ),
+        getColorizedText("Reject", "#E02B00"),
+        declinePendingIntent
+    )
+        .build()
 
     notificationBuilder.addAction(declineAction)
 }
 
-fun addCallAcceptAction(context: Context, notificationBuilder: NotificationCompat.Builder,
-                        callId: String, callType: Int, callInitiatorId: Int, callInitiatorName: String, opponents: ArrayList<Int>) {
+fun addCallAcceptAction(
+    context: Context,
+    notificationBuilder: NotificationCompat.Builder,
+    callId: String,
+    callType: Int,
+    callInitiatorId: Int,
+    callInitiatorName: String,
+    opponents: ArrayList<Int>,
+    userInfo: String
+) {
     val bundle = Bundle()
     bundle.putString(EXTRA_CALL_ID, callId)
     bundle.putInt(EXTRA_CALL_TYPE, callType)
     bundle.putInt(EXTRA_CALL_INITIATOR_ID, callInitiatorId)
     bundle.putString(EXTRA_CALL_INITIATOR_NAME, callInitiatorName)
     bundle.putIntegerArrayList(EXTRA_CALL_OPPONENTS, opponents)
+    bundle.putString(EXTRA_CALL_USER_INFO, userInfo)
 
     val acceptPendingIntent: PendingIntent = PendingIntent.getBroadcast(
-            context,
-            callId.hashCode(),
-            Intent(context, EventReceiver::class.java)
-                    .setAction(ACTION_CALL_ACCEPT)
-                    .putExtras(bundle),
-            PendingIntent.FLAG_UPDATE_CURRENT)
+        context,
+        callId.hashCode(),
+        Intent(context, EventReceiver::class.java)
+            .setAction(ACTION_CALL_ACCEPT)
+            .putExtras(bundle),
+        PendingIntent.FLAG_UPDATE_CURRENT
+    )
     val acceptAction: NotificationCompat.Action = NotificationCompat.Action.Builder(
-            R.drawable.ic_menu_call,
-            getColorizedText("Accept", "#4CB050"),
-            acceptPendingIntent)
-            .build()
+        context.resources.getIdentifier("ic_menu_call", "drawable", context.packageName),
+        getColorizedText("Accept", "#4CB050"),
+        acceptPendingIntent
+    )
+        .build()
     notificationBuilder.addAction(acceptAction)
 }
 
 fun addCallFullScreenIntent(
-        context: Context, notificationBuilder: NotificationCompat.Builder,
-        callId: String, callType: Int, callInitiatorId: Int, callInitiatorName: String, callOpponents: ArrayList<Int>) {
-    val callFullScreenIntent: Intent = createStartIncomingScreenIntent(context, callId, callType, callInitiatorId, callInitiatorName, callOpponents)
+    context: Context,
+    notificationBuilder: NotificationCompat.Builder,
+    callId: String,
+    callType: Int,
+    callInitiatorId: Int,
+    callInitiatorName: String,
+    callOpponents: ArrayList<Int>,
+    userInfo: String
+) {
+    val callFullScreenIntent: Intent = createStartIncomingScreenIntent(
+        context,
+        callId,
+        callType,
+        callInitiatorId,
+        callInitiatorName,
+        callOpponents,
+        userInfo
+    )
     val fullScreenPendingIntent = PendingIntent.getActivity(
-            context,
-            callId.hashCode(),
-            callFullScreenIntent,
-            PendingIntent.FLAG_UPDATE_CURRENT)
+        context,
+        callId.hashCode(),
+        callFullScreenIntent,
+        PendingIntent.FLAG_UPDATE_CURRENT
+    )
     notificationBuilder.setFullScreenIntent(fullScreenPendingIntent, true)
 }
 
-fun addCancelCallNotificationIntent(appContext: Context?, notificationBuilder: NotificationCompat.Builder, callId: String, callType: Int, callInitiatorId: Int, callInitiatorName: String) {
+fun addCancelCallNotificationIntent(
+    appContext: Context?,
+    notificationBuilder: NotificationCompat.Builder,
+    callId: String,
+    callType: Int,
+    callInitiatorId: Int,
+    callInitiatorName: String,
+    userInfo: String
+) {
     val bundle = Bundle()
     bundle.putString(EXTRA_CALL_ID, callId)
     bundle.putInt(EXTRA_CALL_TYPE, callType)
     bundle.putInt(EXTRA_CALL_INITIATOR_ID, callInitiatorId)
     bundle.putString(EXTRA_CALL_INITIATOR_NAME, callInitiatorName)
+    bundle.putString(EXTRA_CALL_USER_INFO, userInfo)
 
     val deleteCallNotificationPendingIntent = PendingIntent.getBroadcast(
-            appContext,
-            callId.hashCode(),
-            Intent(appContext, EventReceiver::class.java)
-                    .setAction(ACTION_CALL_NOTIFICATION_CANCELED)
-                    .putExtras(bundle),
-            PendingIntent.FLAG_UPDATE_CURRENT)
+        appContext,
+        callId.hashCode(),
+        Intent(appContext, EventReceiver::class.java)
+            .setAction(ACTION_CALL_NOTIFICATION_CANCELED)
+            .putExtras(bundle),
+        PendingIntent.FLAG_UPDATE_CURRENT
+    )
     notificationBuilder.setDeleteIntent(deleteCallNotificationPendingIntent)
 }
 
 fun createCallNotificationChannel(notificationManager: NotificationManagerCompat, sound: Uri) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-        val channel = NotificationChannel(CALL_CHANNEL_ID, CALL_CHANNEL_NAME, NotificationManager.IMPORTANCE_HIGH)
-        channel.setSound(sound, AudioAttributes.Builder()
+        val channel = NotificationChannel(
+            CALL_CHANNEL_ID,
+            CALL_CHANNEL_NAME,
+            NotificationManager.IMPORTANCE_HIGH
+        )
+        channel.setSound(
+            sound, AudioAttributes.Builder()
                 .setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
                 .setUsage(AudioAttributes.USAGE_NOTIFICATION_RINGTONE)
-                .build())
+                .build()
+        )
         notificationManager.createNotificationChannel(channel)
     }
 }
 
 fun setNotificationSmallIcon(context: Context, notificationBuilder: NotificationCompat.Builder) {
-    val resID = context.resources.getIdentifier("ic_launcher_foreground", "drawable", context.packageName)
+    val resID =
+        context.resources.getIdentifier("ic_launcher_foreground", "drawable", context.packageName)
     if (resID != 0) {
         notificationBuilder.setSmallIcon(resID)
     } else {
@@ -184,7 +295,11 @@ fun setNotificationSmallIcon(context: Context, notificationBuilder: Notification
 
 fun setNotificationColor(context: Context, notificationBuilder: NotificationCompat.Builder) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-        val accentID = context.resources.getIdentifier("call_notification_color_accent", "color", context.packageName)
+        val accentID = context.resources.getIdentifier(
+            "call_notification_color_accent",
+            "color",
+            context.packageName
+        )
         if (accentID != 0) {
             notificationBuilder.color = context.resources.getColor(accentID, null)
         } else {
